#3398 mv redundant trans to own file
